Updated max_constant stuff to be u128's everywhere.

+++ src/search/mod.rs
@@ -270,7 +270,7 @@ fn find_with_length_and_op<N: Number, const C: usize, V: Verdict<N, C>>(
     thread_id: usize,
     notification_spacing: u128,
     judge: &dyn Fn(&Expression<N, C>) -> V,
-    constant_cap: u8,
+    constant_cap: u128,
     length: usize,
     writer_type: WriterType,
     var_names: Option<&str>,
