+++ src/search/writer/writers/neg_writer.rs
@@ -10,7 +10,7 @@ pub struct NegWriter<N: Number, const C: usize> {
 }
 
 impl<N: Number, const C: usize> NegWriter<N, C> {
-    pub fn new(length: usize, constant_cap: u8) -> Self {
+    pub fn new(length: usize, constant_cap: u128) -> Self {
         Self {
             length,
             child: Box::new(Writer::new(length - 1, WriterContext {location: CHILD_OF_NEG, const_allowed: false}, None, constant_cap)),
